+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### `Added`
 - [PR #60](https://github.com/CDCgov/mira-oxide/pull/60) - Added create_nextflow_samplesheet.rs to replace create_nextflow_samplesheet_i.py and create_nextflow_samplesheet+o.py in MIRA-NF. Takes the input samplesheet and fastq files and puts them into the nextflow sampleseet format.
+- [PR #60](https://github.com/CDCgov/mira-oxide/pull/60) - Added added functionality to create_nextflow_samplesheet.rs that tells user when fastq files are missing or empty
 
 ### `Fixed`
 - [PR #64](https://github.com/CDCgov/mira-oxide/pull/64) - Zoe alignment syntax to be compatible with update